Implementing new SCENE displaying system (1/2)
A new structure for how scenes are displayed is being used, replacing the old "display" file and utilizing Classes/Objects to better organize scenery.

game.state is no longer an attribute, and now the game will fetch scenes directly from a folder as a class, getting access to methods and attributes from that scene alone. On top of this change, almost every display rendering code has been refactored and improved upon.

This initial commit is not finished, and needs further implementation of separating certain variables & getting the game back to a working state.